+++ src/hooks/index.js
@@ -10,4 +10,5 @@ export { default as usePreventScroll } from "./usePreventScroll";
 export { default as useCardFooterButtonClick } from "./useCardFooterButtonClick";
 export { default as useInfiniteScroll } from "./useInfiniteScroll";
 export { default as useProcessCommentsData } from "./useProcessCommentsData";
+export { default as useSubmitComment } from "./useSubmitComment";
 export * from "./post";
